+++ code/lib/core-server/src/server-channel/create-new-story-channel.ts
@@ -2,13 +2,9 @@ import type { Options } from '@storybook/types';
 import type { Channel } from '@storybook/channels';
 import { CREATE_NEW_STORYFILE, CREATE_NEW_STORYFILE_RESULT } from '@storybook/core-events';
 import dedent from 'ts-dedent';
-import { normalizeStories, normalizeStoryPath } from '@storybook/core-common';
-import path from 'path';
-import fs from 'fs';
-import { storyNameFromExport, toId } from '@storybook/csf';
-import slash from 'slash';
-import { userOrAutoTitleFromSpecifier } from '@storybook/preview-api';
+import fs from 'fs/promises';
 import { getNewStoryFile } from '../utils/get-new-story-file';
+import { getStoryId } from '../utils/get-story-id';
 
 interface Data {
   // The filepath of the component for which the Story should be generated for (relative to the project root)
@@ -38,11 +34,11 @@ export function initCreateNewStoryChannel(channel: Channel, options: Options) {
         options
       );
 
-      fs.writeFileSync(storyFilePath, storyFileContent, {
+      await fs.writeFile(storyFilePath, storyFileContent, {
         encoding: 'utf-8',
       });
 
-      const storyId = await getStoryId(options, storyFilePath, exportedStoryName);
+      const storyId = await getStoryId({ storyFilePath, exportedStoryName }, options);
 
       channel.emit(CREATE_NEW_STORYFILE_RESULT, {
         success: true,
@@ -64,40 +60,4 @@ export function initCreateNewStoryChannel(channel: Channel, options: Options) {
   });
 
   return channel;
-}
-
-export async function getStoryId(
-  options: Options,
-  storyFilePath: string,
-  exportedStoryName: string
-) {
-  const stories = await options.presets.apply('stories', [], options);
-
-  const workingDir = process.cwd();
-
-  const normalizedStories = normalizeStories(stories, {
-    configDir: options.configDir,
-    workingDir,
-  });
-
-  const relativePath = path.relative(workingDir, storyFilePath);
-  const importPath = slash(normalizeStoryPath(relativePath));
-
-  const autoTitle = normalizedStories
-    .map((normalizeStory) => userOrAutoTitleFromSpecifier(importPath, normalizeStory))
-    .filter(Boolean)[0];
-
-  if (autoTitle === undefined) {
-    // eslint-disable-next-line local-rules/no-uncategorized-errors
-    throw new Error(dedent`
-    The generation of your new Story file was successful! But it seems that we are unable to index it.
-    Please make sure that the new Story file is matched by the 'stories' glob pattern in your Storybook configuration.
-    The location of the new Story file is: ${relativePath}
-  `);
-  }
-
-  const storyName = storyNameFromExport(exportedStoryName);
-  const storyId = toId(autoTitle as string, storyName);
-
-  return storyId;
 }
