+++ code/lib/core-server/src/utils/get-new-story-file.test.ts
@@ -31,8 +31,9 @@ describe('get-new-story-file', () => {
 
     expect(exportedStoryName).toBe('Default');
     expect(storyFileContent).toMatchInlineSnapshot(`
-      "import { Page } from './Page';
-      import type { Meta, StoryObj } from '@storybook/nextjs';
+      "import type { Meta, StoryObj } from '@storybook/nextjs';
+
+      import { Page } from './Page';
 
       const meta = {
         component: Page
